# Conflicts:
#	src/databricks/labs/ucx/inventory/inventorizer.py
#	src/databricks/labs/ucx/inventory/permissions.py
#	src/databricks/labs/ucx/inventory/permissions_inventory.py
#	src/databricks/labs/ucx/toolkits/group_migration.py
#	tests/unit/test_inventorizer.py
#	tests/unit/test_permissions_inventory.py

+++ tests/integration/test_e2e.py
@@ -1,148 +1,54 @@
 import logging
 import os
 import random
-from typing import Literal
 
-import pytest
 from databricks.sdk import WorkspaceClient
 from databricks.sdk.service import workspace
-from databricks.sdk.service.iam import (
-    AccessControlRequest,
-    AccessControlResponse,
-    Permission,
-    PermissionLevel,
-)
-from pyspark.errors import AnalysisException
+from databricks.sdk.service.iam import PermissionLevel
 
 from databricks.labs.ucx.config import (
     ConnectConfig,
     GroupsConfig,
-    InventoryConfig,
-    InventoryTable,
     MigrationConfig,
     TaclConfig,
 )
-from databricks.labs.ucx.inventory.types import RequestObjectType
-from databricks.labs.ucx.providers.groups_info import GroupMigrationState
+from databricks.labs.ucx.inventory.types import LogicalObjectType, RequestObjectType
 from databricks.labs.ucx.toolkits.group_migration import GroupMigrationToolkit
-
-from .utils import EnvironmentInfo, WorkspaceObjects
 
 logger = logging.getLogger(__name__)
 
 
-def _verify_group_permissions(
-    objects: list | WorkspaceObjects | None,
-    id_attribute: str,
-    request_object_type: RequestObjectType | None,
-    ws: WorkspaceClient,
-    toolkit: GroupMigrationToolkit,
-    target: Literal["backup", "account"],
-):
-    logger.debug(
-        f"Verifying that the permissions of object "
-        f"{request_object_type or id_attribute} were applied to {target} groups"
-    )
-
-    if id_attribute == "workspace_objects":
-        _workspace_objects: WorkspaceObjects = objects
-
-        # list of groups that source the permissions
-        comparison_base = [
-            getattr(mi, "workspace" if target == "backup" else "backup")
-            for mi in toolkit.group_manager.migration_groups_provider.groups
-        ]
-        # list of groups that are the target of the permissions
-        comparison_target = [getattr(mi, target) for mi in toolkit.group_manager.migration_groups_provider.groups]
-
-        root_permissions = ws.permissions.get(
-            request_object_type=RequestObjectType.DIRECTORIES, request_object_id=_workspace_objects.root_dir.object_id
-        )
-        base_group_names = [g.display_name for g in comparison_base]
-        target_group_names = [g.display_name for g in comparison_target]
-
-        base_acls = [a for a in root_permissions.access_control_list if a.group_name in base_group_names]
-
-        target_acls = [a for a in root_permissions.access_control_list if a.group_name in target_group_names]
-
-        assert len(base_acls) == len(target_acls)
-
-    elif id_attribute == "secret_scopes":
-        for scope_name in objects:
-            toolkit.permissions_manager.verify_applied_scope_acls(
-                scope_name, toolkit.group_manager.migration_groups_provider, target
-            )
-
-    elif id_attribute in ("tokens", "passwords"):
-        _typed_objects: list[AccessControlRequest] = objects
-        ws_permissions = [
-            AccessControlResponse(
-                all_permissions=[
-                    Permission(permission_level=o.permission_level, inherited=False, inherited_from_object=None)
-                ],
-                group_name=o.group_name,
-            )
-            for o in _typed_objects
-        ]
-
-        target_permissions = list(
-            filter(
-                lambda p: p.group_name
-                in [getattr(g, target).display_name for g in toolkit.group_manager.migration_groups_provider.groups],
-                ws.permissions.get(
-                    request_object_type=request_object_type, request_object_id=id_attribute
-                ).access_control_list,
-            )
-        )
-
-        sorted_ws = sorted(ws_permissions, key=lambda p: p.group_name)
-        sorted_target = sorted(target_permissions, key=lambda p: p.group_name)
-
-        assert [p.all_permissions for p in sorted_ws] == [p.all_permissions for p in sorted_target]
-    else:
-        for _object in objects:
-            toolkit.permissions_manager.verify_applied_permissions(
-                request_object_type,
-                getattr(_object, id_attribute),
-                toolkit.group_manager.migration_groups_provider,
-                target,
-            )
-
-
-def _verify_roles_and_entitlements(
-    migration_state: GroupMigrationState,
-    ws: WorkspaceClient,
-    target: Literal["backup", "account"],
-):
-    for el in migration_state.groups:
-        comparison_base = getattr(el, "workspace" if target == "backup" else "backup")
-        comparison_target = getattr(el, target)
-
-        base_group_info = ws.groups.get(comparison_base.id)
-        target_group_info = ws.groups.get(comparison_target.id)
-
-        assert base_group_info.roles == target_group_info.roles
-        assert base_group_info.entitlements == target_group_info.entitlements
-
-
 def test_e2e(
-    env: EnvironmentInfo,
-    inventory_table: InventoryTable,
     ws: WorkspaceClient,
-    verifiable_objects: list[tuple[list, str, RequestObjectType | None]],
+    make_schema,
+    make_ucx_group,
     make_instance_pool,
     make_instance_pool_permissions,
     make_cluster,
     make_cluster_permissions,
+    make_cluster_policy,
+    make_cluster_policy_permissions,
+    make_model,
+    make_registered_model_permissions,
+    make_experiment,
+    make_experiment_permissions,
     make_job,
     make_job_permissions,
+    make_notebook,
+    make_notebook_permissions,
+    make_directory,
+    make_directory_permissions,
     make_pipeline,
     make_pipeline_permissions,
     make_secret_scope,
     make_secret_scope_acl,
+    make_authorization_permissions,
+    make_warehouse,
+    make_warehouse_permissions,
 ):
-    logger.debug(f"Test environment: {env.test_uid}")
-    ws_group = env.groups[0][0]
+    ws_group, acc_group = make_ucx_group()
+
+    to_verify = set()
 
     pool = make_instance_pool()
     make_instance_pool_permissions(
@@ -150,7 +56,7 @@ def test_e2e(
         permission_level=random.choice([PermissionLevel.CAN_ATTACH_TO, PermissionLevel.CAN_MANAGE]),
         group_name=ws_group.display_name,
     )
-    verifiable_objects.append(([pool], "instance_pool_id", RequestObjectType.INSTANCE_POOLS))
+    to_verify.add((RequestObjectType.INSTANCE_POOLS, pool.instance_pool_id))
 
     cluster = make_cluster(instance_pool_id=os.environ["TEST_INSTANCE_POOL_ID"], single_node=True)
     make_cluster_permissions(
@@ -160,9 +66,60 @@ def test_e2e(
         ),
         group_name=ws_group.display_name,
     )
-    verifiable_objects.append(
-        ([cluster], "cluster_id", RequestObjectType.CLUSTERS),
+    to_verify.add((RequestObjectType.CLUSTERS, cluster.cluster_id))
+
+    cluster_policy = make_cluster_policy()
+    make_cluster_policy_permissions(
+        object_id=cluster_policy.policy_id,
+        permission_level=random.choice([PermissionLevel.CAN_USE]),
+        group_name=ws_group.display_name,
     )
+    to_verify.add((RequestObjectType.CLUSTER_POLICIES, cluster_policy.policy_id))
+
+    model = make_model()
+    make_registered_model_permissions(
+        object_id=model.id,
+        permission_level=random.choice(
+            [
+                PermissionLevel.CAN_READ,
+                PermissionLevel.CAN_MANAGE,
+                PermissionLevel.CAN_MANAGE_PRODUCTION_VERSIONS,
+                PermissionLevel.CAN_MANAGE_STAGING_VERSIONS,
+            ]
+        ),
+        group_name=ws_group.display_name,
+    )
+    to_verify.add((RequestObjectType.REGISTERED_MODELS, model.id))
+
+    experiment = make_experiment()
+    make_experiment_permissions(
+        object_id=experiment.experiment_id,
+        permission_level=random.choice(
+            [PermissionLevel.CAN_MANAGE, PermissionLevel.CAN_READ, PermissionLevel.CAN_EDIT]
+        ),
+        group_name=ws_group.display_name,
+    )
+    to_verify.add((RequestObjectType.EXPERIMENTS, experiment.experiment_id))
+
+    directory = make_directory()
+    make_directory_permissions(
+        object_id=directory,
+        permission_level=random.choice(
+            [PermissionLevel.CAN_READ, PermissionLevel.CAN_MANAGE, PermissionLevel.CAN_EDIT, PermissionLevel.CAN_RUN]
+        ),
+        group_name=ws_group.display_name,
+    )
+    to_verify.add((RequestObjectType.DIRECTORIES, ws.workspace.get_status(directory).object_id))
+
+    notebook = make_notebook(path=f"{directory}/sample.py")
+    make_notebook_permissions(
+        object_id=notebook,
+        permission_level=random.choice(
+            [PermissionLevel.CAN_READ, PermissionLevel.CAN_MANAGE, PermissionLevel.CAN_EDIT, PermissionLevel.CAN_RUN]
+        ),
+        group_name=ws_group.display_name,
+    )
+    to_verify.add((RequestObjectType.NOTEBOOKS, ws.workspace.get_status(notebook).object_id))
 
     job = make_job()
     make_job_permissions(
@@ -172,9 +129,7 @@ def test_e2e(
         ),
         group_name=ws_group.display_name,
     )
-    verifiable_objects.append(
-        ([job], "job_id", RequestObjectType.JOBS),
-    )
+    to_verify.add((RequestObjectType.JOBS, job.job_id))
 
     pipeline = make_pipeline()
     make_pipeline_permissions(
@@ -182,25 +137,39 @@ def test_e2e(
         permission_level=random.choice([PermissionLevel.CAN_VIEW, PermissionLevel.CAN_RUN, PermissionLevel.CAN_MANAGE]),
         group_name=ws_group.display_name,
     )
-    verifiable_objects.append(
-        ([pipeline], "pipeline_id", RequestObjectType.PIPELINES),
-    )
+    to_verify.add((RequestObjectType.PIPELINES, pipeline.pipeline_id))
 
     scope = make_secret_scope()
     make_secret_scope_acl(scope=scope, principal=ws_group.display_name, permission=workspace.AclPermission.WRITE)
-    verifiable_objects.append(([scope], "secret_scopes", None))
+    to_verify.add((LogicalObjectType.SECRET_SCOPE, scope))
+
+    make_authorization_permissions(
+        object_id="tokens",
+        permission_level=PermissionLevel.CAN_USE,
+        group_name=ws_group.display_name,
+    )
+    to_verify.add((RequestObjectType.AUTHORIZATION, "tokens"))
+
+    warehouse = make_warehouse()
+    make_warehouse_permissions(
+        object_id=warehouse.id,
+        permission_level=random.choice([PermissionLevel.CAN_USE, PermissionLevel.CAN_MANAGE]),
+        group_name=ws_group.display_name,
+    )
+    to_verify.add((RequestObjectType.SQL_WAREHOUSES, warehouse.id))
 
     config = MigrationConfig(
         connect=ConnectConfig.from_databricks_config(ws.config),
-        inventory=InventoryConfig(table=inventory_table),
-        groups=GroupsConfig(selected=[g[0].display_name for g in env.groups]),
+        inventory_database=make_schema(catalog="hive_metastore").split(".")[-1],
+        groups=GroupsConfig(selected=[ws_group.display_name]),
+        workspace_start_path=directory,
         tacl=TaclConfig(auto=True),
         log_level="DEBUG",
     )
     toolkit = GroupMigrationToolkit(config)
     toolkit.prepare_environment()
 
-    group_migration_state = toolkit.group_manager.migration_groups_provider
+    group_migration_state = toolkit._group_manager.migration_groups_provider
     for _info in group_migration_state.groups:
         _ws = ws.groups.get(id=_info.workspace.id)
         _backup = ws.groups.get(id=_info.backup.id)
@@ -212,17 +181,11 @@ def test_e2e(
 
     toolkit.cleanup_inventory_table()
 
-    with pytest.raises(AnalysisException):
-        toolkit.table_manager.spark.catalog.getTable(toolkit.table_manager.config.table.to_spark())
-
     toolkit.inventorize_permissions()
 
     toolkit.apply_permissions_to_backup_groups()
 
-    for _objects, id_attribute, request_object_type in verifiable_objects:
-        _verify_group_permissions(_objects, id_attribute, request_object_type, ws, toolkit, "backup")
-
-    _verify_roles_and_entitlements(group_migration_state, ws, "backup")
+    toolkit.verify_permissions_on_backup_groups(to_verify)
 
     toolkit.replace_workspace_groups_with_account_groups()
 
@@ -234,10 +197,7 @@ def test_e2e(
 
     toolkit.apply_permissions_to_account_groups()
 
-    for _objects, id_attribute, request_object_type in verifiable_objects:
-        _verify_group_permissions(_objects, id_attribute, request_object_type, ws, toolkit, "account")
-
-    _verify_roles_and_entitlements(group_migration_state, ws, "account")
+    toolkit.verify_permissions_on_account_groups(to_verify)
 
     toolkit.delete_backup_groups()
